seperate lua logic better
