+++ hitokage-core/src/lib.rs
@@ -1,1 +1,4 @@
 
+pub const HITOKAGE_STATUSBAR_HEIGHT: i32 = 64;
+pub mod bar;
+pub mod common;
