+++ hitokage-lua/src/bar.rs
@@ -1,0 +1,89 @@
+use hitokage_core::bar::BarProps;
+use mlua::{Lua, LuaSerdeExt, Value};
+use relm4::{Component, ComponentSender};
+
+pub fn make<'lua, C>(lua: &'lua Lua, sender: &ComponentSender<C>) -> anyhow::Result<mlua::Table<'lua>>
+where
+  C: Component<Input = crate::AppMsg>,
+  <C as Component>::Output: std::marker::Send,
+{
+  let table = lua.create_table()?;
+
+  {
+    table.set(
+      "create",
+      lua.create_function({
+        let sender = sender.clone();
+        move |lua_inner, value: Value| {
+          let props: BarProps = lua_inner.from_value(value)?;
+          sender.input(<C as Component>::Input::LuaHook(crate::LuaHook {
+            t: crate::LuaHookType::CreateBar(props),
+            callback: Box::new(|_| Ok(())),
+          }));
+          Ok(())
+        }
+      })?,
+    )?;
+  }
+
+  Ok(table)
+}
+
+#[cfg(test)]
+mod tests {
+  use crate::{assert_lua_type, AppMsg};
+  use gtk4::prelude::*;
+  use mlua::{Lua, Table, Value};
+  use relm4::prelude::*;
+  use relm4::{ComponentParts, ComponentSender, SimpleComponent};
+
+  struct DummyComponent {}
+
+  #[relm4::component]
+  impl SimpleComponent for DummyComponent {
+    type Input = crate::AppMsg;
+    type Output = ();
+    type Init = ();
+
+    view! {
+      gtk::ApplicationWindow {
+        set_visible: false,
+      },
+    }
+
+    fn init(_: Self::Init, root: Self::Root, sender: ComponentSender<Self>) -> ComponentParts<Self> {
+      let model = DummyComponent {};
+
+      let tests = || -> anyhow::Result<()> {
+        let lua = Lua::new();
+        let table = super::make(&lua, &sender)?;
+        lua.globals().set("table", table)?;
+        let value: Value = lua.globals().get("table")?;
+
+        let table: Table = assert_lua_type!(value, mlua::Table);
+        assert_lua_type!(table.get::<&str, Value>("create")?, mlua::Function);
+        assert_eq!(table.len()?, 0);
+        assert_eq!(table.pairs::<String, Value>().count(), 1);
+
+        relm4::main_application().quit();
+
+        Ok(())
+      };
+      let _ = tests().unwrap();
+
+      let widgets = view_output!();
+
+      ComponentParts { model, widgets }
+    }
+
+    fn update(&mut self, _: AppMsg, _: ComponentSender<Self>) {}
+  }
+
+  #[test]
+  fn test_all() -> anyhow::Result<()> {
+    let app = RelmApp::new("com.example.hitokagetest");
+    app.run::<DummyComponent>(());
+
+    Ok(())
+  }
+}
