+++ hitokage/src/komorebihelpers.rs
@@ -1,7 +1,7 @@
 
 // reserves the space for the bar
 fn reserve_space(props: &Bar) -> () {
-
+  
 
   ()
 }
